chore(web): requireable module for common CI test-config
